chore: support esm and cjs module system
